+++ app-backend/batch/src/main/scala/notification/NotifyIngestStatus.scala
@@ -3,11 +3,8 @@ package com.rasterfoundry.batch.notification
 import com.rasterfoundry.batch.Job
 import com.rasterfoundry.batch.util.conf.Config
 import com.rasterfoundry.common.RollbarNotifier
-import com.rasterfoundry.common.notification.Email.{
-  EmailConfig,
-  NotificationEmail
-}
 import com.rasterfoundry.database.filter.Filterables._
+import com.rasterfoundry.database.notification.Notify
 import com.rasterfoundry.database.util.RFTransactor
 import com.rasterfoundry.database.{
   PlatformDao,
@@ -16,6 +13,7 @@ import com.rasterfoundry.database.{
   SceneDao
 }
 import com.rasterfoundry.datamodel._
+import com.rasterfoundry.notification.email.Model._
 
 import cats.effect.IO
 import cats.implicits._
@@ -23,13 +21,12 @@ import doobie._
 import doobie.implicits._
 import doobie.postgres.implicits._
 import doobie.util.transactor.Transactor
-import org.apache.commons.mail.Email
 
 import java.util.UUID
 
 final case class NotifyIngestStatus(sceneId: UUID)(
-    implicit val xa: Transactor[IO])
-    extends Config
+    implicit val xa: Transactor[IO]
+) extends Config
     with RollbarNotifier {
 
   val name = NotifyIngestStatus.name
@@ -42,7 +39,8 @@ final case class NotifyIngestStatus(sceneId: UUID)(
     for {
       layers <- ProjectLayerDao.query
         .filter(
-          fr"id IN (SELECT project_layer_id FROM scenes_to_layers WHERE scene_id = ${sceneId})")
+          fr"id IN (SELECT project_layer_id FROM scenes_to_layers WHERE scene_id = ${sceneId})"
+        )
         .list
       projects <- layers.filter(!_.projectId.isEmpty).toNel match {
         case Some(nel) =>
@@ -65,14 +63,17 @@ final case class NotifyIngestStatus(sceneId: UUID)(
   def createIngestEmailContentForConsumers(
       pU: PlatformWithUsersSceneProjects,
       scene: Scene,
-      ingestStatus: String): (String, String, String) = {
+      ingestStatus: String
+  ): (Subject, HtmlBody, PlainBody) = {
     val platformHost =
       pU.pubSettings.platformHost.getOrElse("app.rasterfoundry.com")
     ingestStatus match {
       case status: String if status == "INGESTED" =>
         (
-          s"${pU.platName}: Your ${pU.projectName}'s scene ${scene.name} is ready to view",
-          s"""
+          Subject(
+            s"${pU.platName}: Your ${pU.projectName}'s scene ${scene.name} is ready to view"
+          ),
+          HtmlBody(s"""
           <html>
             <p>${pU.uName},</p><br>
             <p>The scene "${scene.name}" has been successfully ingested into your project: ${pU.projectName}! You can access
@@ -81,20 +82,22 @@ final case class NotifyIngestStatus(sceneId: UUID)(
             <p>If you have questions, support is available via in-app chat at ${platformHost} or less quickly via email to ${pU.pubSettings.emailSupport}.</p>
             <p>- The ${pU.platName} Team</p>
           </html>
-          """,
-          s"""
+          """),
+          PlainBody(s"""
              | ${pU.uName}: The scene "${scene.name}" has been successfully ingested into your project: ${pU.projectName}!
              | You can access this project at here: https://${platformHost}/projects/edit/${pU.projectId}/scenes or
              | any past projects you've created at any time here: https://${platformHost}/projects/ . If you have
              | questions, support is available via in-app chat at ${platformHost} or less quickly via email to
              | ${pU.pubSettings.emailSupport}.
              | - The ${pU.platName} Team
-          """.trim.stripMargin
+          """.trim.stripMargin)
         )
       case status: String if status == "FAILED" =>
         (
-          s"${pU.platName}: Your ${pU.projectName}'s scene ${scene.name} failed to ingest",
-          s"""
+          Subject(
+            s"${pU.platName}: Your ${pU.projectName}'s scene ${scene.name} failed to ingest"
+          ),
+          HtmlBody(s"""
           <html>
             <p>${pU.uName},</p><br>
             <p>The scene "${scene.name}" in your project: ${pU.projectName} has failed to ingest. But you can access
@@ -103,14 +106,16 @@ final case class NotifyIngestStatus(sceneId: UUID)(
             <p>If you have questions, support is available via in-app chat at ${platformHost} or less quickly via email to ${pU.pubSettings.emailSupport}.</p>
             <p>- The ${pU.platName} Team</p>
           </html>
-          """,
-          s"""
+          """),
+          PlainBody(
+            s"""
              | ${pU.uName}: The scene "${scene.name}" in your project: ${pU.projectName} has failed to ingest. But you can
              | access this project at here: https://${platformHost}/projects/edit/${pU.projectId}/scenes or
              | any past projects you've created at any time here: https://${platformHost}/projects/ . If you have
              | questions, support is available via in-app chat at ${platformHost} or less quickly via email to ${pU.pubSettings.emailSupport}.
              | - The ${pU.platName} Team
           """.trim.stripMargin
+          )
         )
     }
   }
@@ -118,44 +123,47 @@ final case class NotifyIngestStatus(sceneId: UUID)(
   def createIngestEmailContentForOwner(
       pO: PlatformWithSceneOwner,
       scene: Scene,
-      ingestStatus: String): (String, String, String) = {
+      ingestStatus: String
+  ): (Subject, HtmlBody, PlainBody) = {
     val platformHost =
       pO.pubSettings.platformHost.getOrElse("app.rasterfoundry.com")
     ingestStatus match {
       case status: String if status == "INGESTED" =>
         (
-          s"${pO.platName}: Your scene ${scene.name} is ready to view",
-          s"""
+          Subject(s"${pO.platName}: Your scene ${scene.name} is ready to view"),
+          HtmlBody(s"""
           <html>
             <p>${pO.uName},</p><br>
             <p>The scene "${scene.name}" has been successfully ingested!</p>
             <p>If you have questions, support is available via in-app chat at ${platformHost} or less quickly via email to ${pO.pubSettings.emailSupport}.</p>
             <p>- The ${pO.platName} Team</p>
           </html>
-          """,
-          s"""
+          """),
+          PlainBody(
+            s"""
              | ${pO.uName}: The scene "${scene.name}" has been successfully ingested!
              | If you have questions, support is available via in-app chat at ${platformHost} or less quickly via email to ${pO.pubSettings.emailSupport}.
              | - The ${pO.platName} Team
           """.trim.stripMargin
+          )
         )
       case status: String if status == "FAILED" =>
         (
-          s"${pO.platName}: Your scene ${scene.name} failed to ingest",
-          s"""
+          Subject(s"${pO.platName}: Your scene ${scene.name} failed to ingest"),
+          HtmlBody(s"""
           <html>
             <p>${pO.uName},</p><br>
             <p>The scene "${scene.name}" has failed to ingest. </p>
             <p>If you have questions, support is available via in-app chat at ${platformHost} or less quickly via email to ${pO.pubSettings.emailSupport}.</p>
             <p>- The ${pO.platName} Team</p>
           </html>
-          """,
-          s"""
+          """),
+          PlainBody(s"""
              | ${pO.uName}: The scene "${scene.name}" has failed to ingest.
              | If you have questions,
              | support is available via in-app chat at ${platformHost} or less quickly via email to ${pO.pubSettings.emailSupport}.
              | - The ${pO.platName} Team
-          """.trim.stripMargin
+          """.trim.stripMargin)
         )
     }
   }
@@ -163,110 +171,63 @@ final case class NotifyIngestStatus(sceneId: UUID)(
   def sendIngestStatusEmailToConsumers(
       platformsWithConsumers: List[PlatformWithUsersSceneProjects],
       scene: Scene,
-      ingestStatus: String) =
-    platformsWithConsumers.map(pU => {
-      val email = new NotificationEmail
-
+      ingestStatus: String
+  ): IO[Unit] =
+    (platformsWithConsumers traverse { pU =>
       (pU.getUserEmail, pU.pubSettings.emailIngestNotification) match {
-        case ("", true) =>
-          logger.warn(email.userEmailNotificationDisabledWarning(pU.uId))
-        case ("", false) =>
-          logger.warn(
-            email.userEmailNotificationDisabledWarning(pU.uId) ++ " " ++ email
-              .platformNotSubscribedWarning(pU.platId.toString()))
         case (userEmailAddress, true) =>
-          (pU.pubSettings.emailSmtpHost,
-           pU.pubSettings.emailSmtpPort,
-           pU.pubSettings.emailSmtpEncryption,
-           pU.pubSettings.emailSmtpUserName,
-           pU.priSettings.emailPassword,
-           userEmailAddress) match {
-            case (host: String,
-                  port: Int,
-                  encryption: String,
-                  platSmtpUserName: String,
-                  pw: String,
-                  userEmail: String)
-                if email.isValidEmailSettings(host,
-                                              port,
-                                              encryption,
-                                              platSmtpUserName,
-                                              pw,
-                                              userEmail) =>
-              val (ingestEmailSubject, htmlBody, plainBody) =
-                createIngestEmailContentForConsumers(pU, scene, ingestStatus)
-              email
-                .setEmail(
-                  EmailConfig(host, port, encryption, platSmtpUserName, pw),
-                  userEmail,
-                  ingestEmailSubject,
-                  htmlBody,
-                  plainBody,
-                  pU.pubSettings.emailFrom,
-                  pU.pubSettings.emailFromDisplayName
-                )
-                .map((configuredEmail: Email) => configuredEmail.send)
-              logger.info(s"Notified project owner ${pU.uId}.")
-            case _ =>
-              logger.warn(
-                email.insufficientSettingsWarning(pU.platId.toString(), pU.uId))
+          val (ingestEmailSubject, htmlBody, plainBody) =
+            createIngestEmailContentForConsumers(pU, scene, ingestStatus)
+          Notify.sendEmail(
+            pU.pubSettings,
+            pU.priSettings,
+            userEmailAddress,
+            ingestEmailSubject.underlying,
+            htmlBody.underlying,
+            plainBody.underlying
+          ) map { _ =>
+            logger.info(s"Notified project owner ${pU.uId}.")
           }
         case (_, false) =>
-          logger.warn(email.platformNotSubscribedWarning(pU.platId.toString()))
-      }
-    })
-
-  def sendIngestStatusEmailToOwner(pO: PlatformWithSceneOwner,
-                                   scene: Scene,
-                                   ingestStatus: String) = {
-    val email = new NotificationEmail
-
-    (pO.getUserEmail, pO.pubSettings.emailIngestNotification) match {
-      case ("", true) =>
-        logger.warn(email.userEmailNotificationDisabledWarning(pO.uId))
-      case ("", false) =>
-        logger.warn(
-          email.userEmailNotificationDisabledWarning(pO.uId) ++ " " ++ email
-            .platformNotSubscribedWarning(pO.platId.toString()))
-      case (userEmailAddress, true) =>
-        (pO.pubSettings.emailSmtpHost,
-         pO.pubSettings.emailSmtpPort,
-         pO.pubSettings.emailSmtpEncryption,
-         pO.pubSettings.emailSmtpUserName,
-         pO.priSettings.emailPassword,
-         userEmailAddress) match {
-          case (host: String,
-                port: Int,
-                encryption: String,
-                platSmtpUserName: String,
-                pw: String,
-                userEmail: String)
-              if email.isValidEmailSettings(host,
-                                            port,
-                                            encryption,
-                                            platSmtpUserName,
-                                            pw,
-                                            userEmail) =>
-            val (ingestEmailSubject, htmlBody, plainBody) =
-              createIngestEmailContentForOwner(pO, scene, ingestStatus)
-            email
-              .setEmail(
-                EmailConfig(host, port, encryption, platSmtpUserName, pw),
-                userEmail,
-                ingestEmailSubject,
-                htmlBody,
-                plainBody,
-                pO.pubSettings.emailFrom,
-                pO.pubSettings.emailFromDisplayName
-              )
-              .map((configuredEmail: Email) => configuredEmail.send)
-            logger.info(s"Notified scene owner ${pO.uId}.")
-          case _ =>
+          IO {
             logger.warn(
-              email.insufficientSettingsWarning(pO.platId.toString(), pO.uId))
+              s"platform ${pU.platId} has not enabled email notifications"
+            )
+          }
+      }
+    }).void
+
+  def sendIngestStatusEmailToOwner(
+      pO: PlatformWithSceneOwner,
+      scene: Scene,
+      ingestStatus: String
+  ): IO[Unit] = {
+    (pO.getUserEmail, pO.pubSettings.emailIngestNotification) match {
+      case ("", _) =>
+        IO {
+          logger.warn(
+            s"User ${pO.uId} disabled email notifications."
+          )
+        }
+      case (userEmailAddress, true) =>
+        val (ingestEmailSubject, htmlBody, plainBody) =
+          createIngestEmailContentForOwner(pO, scene, ingestStatus)
+        Notify.sendEmail(
+          pO.pubSettings,
+          pO.priSettings,
+          userEmailAddress,
+          ingestEmailSubject.underlying,
+          htmlBody.underlying,
+          plainBody.underlying
+        ) map { _ =>
+          logger.info(s"Notified project owner ${pO.uId}.")
         }
       case (_, false) =>
-        logger.warn(email.platformNotSubscribedWarning(pO.platId.toString()))
+        IO {
+          logger.warn(
+            s"platform ${pO.platId} has not enabled email notifications"
+          )
+        }
     }
   }
 
@@ -282,20 +243,25 @@ final case class NotifyIngestStatus(sceneId: UUID)(
           .getPlatUsersAndProjByConsumerAndSceneID(consumerIds, sceneId)
           .transact(xa)
           .unsafeRunSync()
-        sendIngestStatusEmailToConsumers(platformsWithConsumers,
-                                         scene,
-                                         ingestStatus)
+        sendIngestStatusEmailToConsumers(
+          platformsWithConsumers,
+          scene,
+          ingestStatus
+        )
       case _ => logger.warn(s"Scene ${sceneId} is not in any project yet.")
     }
     ()
   }
 
-  def notifyOwners(scene: Scene, ingestStatus: String): Unit = {
+  def notifyOwners(scene: Scene, ingestStatus: String): IO[Unit] = {
     logger.info("Notifying owner...")
 
     if (scene.owner == auth0Config.systemUser) {
-      logger.warn(
-        s"Owner of scene ${sceneId} is a system user. Email not sent.")
+      IO {
+        logger.warn(
+          s"Owner of scene ${sceneId} is a system user. Email not sent."
+        )
+      }
     } else {
       val platformsWithSceneOwner = PlatformDao
         .getPlatAndUsersBySceneOwnerId(scene.owner)
@@ -308,7 +274,8 @@ final case class NotifyIngestStatus(sceneId: UUID)(
   def run(): Unit = {
 
     logger.info(
-      s"Notifying owner and consumer about ingest status for scene ${sceneId}...")
+      s"Notifying owner and consumer about ingest status for scene ${sceneId}..."
+    )
 
     val sceneIO = for {
       sceneO <- SceneDao.getSceneById(sceneId)
@@ -325,7 +292,8 @@ final case class NotifyIngestStatus(sceneId: UUID)(
             notifyConsumers(scene, ingestStatus)
           case _ =>
             logger.warn(
-              s"Won't send an email unless the scene ${sceneId} is ingested or failed.")
+              s"Won't send an email unless the scene ${sceneId} is ingested or failed."
+            )
         }
       case _ => logger.warn(s"No matched scene of id: ${sceneId}")
     }